requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
tabulate>=0.9.0